backend: refactor: rename backupAirtable.js to airtableBackup.js (verb phrase to noun phrase) to indicate that this file is not a runnable script
